requirements.txt
fastapi
uvicorn[standard]
sqlmodel
opencv-python
pillow
requests
python-dotenv
dashscope